Fix Filter and Search By JBs, Fix Scroll Hor And Vertical for Detail JB/TSD Screen In GrapperScanScene

+++ TTC_AR/Assets/Scripts/Dropdown_On_ValueChange.cs
@@ -1,16 +1,11 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using UnityEngine;
 using UnityEngine.UI;
 using TMPro;
-using UnityEngine.AddressableAssets;
-using UnityEngine.ResourceManagement.AsyncOperations;
-using UnityEngine.Networking;
-using Newtonsoft.Json;
-using System;
 using EasyUI.Progress;
-using System.Threading;
 
 public class Dropdown_On_ValueChange : MonoBehaviour
 {
@@ -18,6 +13,7 @@ public class Dropdown_On_ValueChange : MonoBehaviour
     public SearchableDropDown searchableDropDown;
     public Open_Detail_Image open_Detail_Image;
     public EventPublisher eventPublisher;
+
     [SerializeField] private RectTransform contentTransform;
     [SerializeField] private TMP_Text code_Value_Text, function_Value_Text, range_Value_Text, io_Value_Text, jb_Connection_Value_Text, jb_Connection_Location_Text;
     [SerializeField] private Image JB_Location_Image_Prefab, JB_Connection_Wiring_Image_Prefab;
@@ -25,44 +21,37 @@ public class Dropdown_On_ValueChange : MonoBehaviour
     [SerializeField] private ScrollRect scrollRect;
     [SerializeField] private Dictionary<string, Sprite> spriteCache = new();
     [SerializeField] private List<Image> instantiatedImages = new();
-    private string currentLoadedDeviceCode;
+    [SerializeField] private Transform deviceInfo;
+
+    private string current_Search_Content = "";
     private Dictionary<string, DeviceInformationModel> deviceDictionary;
+    private Dictionary<string, JBInformationModel> jBDictionary;
+
     private string _jbName = "";
     private string _moduleName = "";
+
     private void Awake()
     {
-        Debug.Log("Dropdown_On_ValueChange Awake");
-        searchableDropDown ??= searchableDropDown.GetComponent<SearchableDropDown>();
+        searchableDropDown ??= GameObject.Find("Searchable").GetComponent<SearchableDropDown>();
     }
+
     private void OnEnable()
     {
-        // Đăng ký lắng nghe sự kiện
-        if (eventPublisher != null)
-        {
-            eventPublisher.OnOrientationChanged += HandleOrientationChange;
-        }
+        // if (eventPublisher != null)
+        // {
+        //     eventPublisher.OnOrientationChanged += HandleOrientationChange;
+        // }
     }
+
     private void OnDisable()
     {
-        // Hủy đăng ký sự kiện khi đối tượng bị disable
-        if (eventPublisher != null)
-        {
-            eventPublisher.OnOrientationChanged -= HandleOrientationChange;
-        }
-        ResetResources();
+        // if (eventPublisher != null)
+        // {
+        //     eventPublisher.OnOrientationChanged -= HandleOrientationChange;
+        // }
+        // ResetResources();
     }
-    // Hàm xử lý khi sự kiện được kích hoạt
-    private void HandleOrientationChange(ScreenOrientation newOrientation)
-    {
-        if (newOrientation == ScreenOrientation.Portrait)
-        {
-            bottom_App_Bar.SetActive(true);
-        }
-        else if (newOrientation == ScreenOrientation.LandscapeLeft || newOrientation == ScreenOrientation.LandscapeRight)
-        {
-            bottom_App_Bar.SetActive(false);
-        }
-    }
+
     private void Start()
     {
         LoadData();
@@ -72,27 +61,26 @@ public class Dropdown_On_ValueChange : MonoBehaviour
     {
         try
         {
-            InitilizeUIElements();
+            InitUIElements();
             Prepare_Device_Dictionary_For_Searching();
+            Prepare_JB_Dictionary_For_Searching();
             searchableDropDown.Initialize();
             searchableDropDown.inputField.onValueChanged.AddListener(OnInputValueChanged);
-            if (string.IsNullOrEmpty(searchableDropDown.inputField.text))
-            {
-                searchableDropDown.Set_Initial_Text_Field_Value();
-            }
+            searchableDropDown.SetInitialTextFieldValue();
         }
         catch (Exception e)
         {
-            Debug.LogWarning(e.Message);
+            Debug.LogWarning(e.Message + " " + e.StackTrace + " " + e.Source + " " + e.InnerException);
         }
     }
-    private void InitilizeUIElements()
+
+    private void InitUIElements()
     {
         var content = prefab_Device.transform.Find("Content");
         scrollRect ??= prefab_Device.GetComponent<ScrollRect>();
         contentTransform ??= content.GetComponent<RectTransform>();
 
-        var deviceInfo = content.Find("Device_information");
+        deviceInfo = content.Find("Device_information");
         code_Value_Text ??= deviceInfo.Find("Code_group/Code_value").GetComponent<TMP_Text>();
         function_Value_Text ??= deviceInfo.Find("Function_group/Function_value").GetComponent<TMP_Text>();
         range_Value_Text ??= deviceInfo.Find("Range_group/Range_value").GetComponent<TMP_Text>();
@@ -101,45 +89,87 @@ public class Dropdown_On_ValueChange : MonoBehaviour
         var jbConnectionGroup = content.Find("JB_Connection_group/JB_Connection_text_group");
         jb_Connection_Value_Text ??= jbConnectionGroup.Find("JB_Connection_value").GetComponent<TMP_Text>();
         jb_Connection_Location_Text ??= jbConnectionGroup.Find("JB_Connection_location").GetComponent<TMP_Text>();
-        //   module_Image ??= content.Find("Module_group/Real_Module_Image").GetComponent<Image>();
+
         JB_Connection_Group ??= content.Find("JB_Connection_group").gameObject;
         JB_Location_Image_Prefab ??= JB_Connection_Group.transform.Find("JB_Location_Image").GetComponent<Image>();
         JB_Connection_Wiring_Image_Prefab ??= JB_Connection_Group.transform.Find("JB_Connection_Wiring").GetComponent<Image>();
     }
     private void Prepare_Device_Dictionary_For_Searching()
     {
-        deviceDictionary = GlobalVariable_Search_Devices.temp_ListDeviceInformationModel.ToDictionary(d => d.Code);
-        var functionDictionary = GlobalVariable_Search_Devices.temp_ListDeviceInformationModel.ToDictionary(d => d.Function);
-        foreach (var keyValuePair in functionDictionary)
+        var tempListDevice = GlobalVariable_Search_Devices.temp_ListDeviceInformationModel;
+
+        deviceDictionary = tempListDevice
+            .GroupBy(device => device.Code.ToLower())
+            .ToDictionary(g => g.Key, g => g.First());
+
+        foreach (var device in tempListDevice)
         {
-            if (deviceDictionary.ContainsKey(keyValuePair.Key))
+            var functionKey = device.Function.ToLower();
+            if (!deviceDictionary.ContainsKey(functionKey))
             {
-                deviceDictionary[keyValuePair.Key] = keyValuePair.Value;
-            }
-            else
-            {
-                deviceDictionary.Add(keyValuePair.Key, keyValuePair.Value);
+                deviceDictionary.Add(functionKey, device);
             }
         }
     }
 
+    private void Prepare_JB_Dictionary_For_Searching()
+    {
+        var tempListJB = GlobalVariable_Search_Devices.temp_ListJBInformationModel;
+
+        jBDictionary = tempListJB
+            .GroupBy(jb => jb.Name.ToLower())
+            .ToDictionary(g => g.Key, g => g.First());
+    }
+    // private void Prepare_Device_Dictionary_For_Searching()
+    // {
+    //     deviceDictionary = GlobalVariable_Search_Devices.temp_ListDeviceInformationModel.ToDictionary(d => d.Code);
+    //     var functionDictionary = GlobalVariable_Search_Devices.temp_ListDeviceInformationModel.ToDictionary(d => d.Function);
+    //     foreach (var keyValuePair in functionDictionary)
+    //     {
+    //         if (deviceDictionary.ContainsKey(keyValuePair.Key))
+    //         {
+    //             deviceDictionary[keyValuePair.Key] = keyValuePair.Value;
+    //         }
+    //         else
+    //         {
+    //             deviceDictionary.Add(keyValuePair.Key, keyValuePair.Value);
+    //         }
+    //     }
+    // }
     private void OnInputValueChanged(string input)
     {
-        if (input == currentLoadedDeviceCode)
+        if (input.ToLower() == current_Search_Content.ToLower())
         {
             return;
         }
-        if (!deviceDictionary.TryGetValue(input, out var device))
+        Debug.Log(searchableDropDown.filter_Type);
+        if (searchableDropDown.filter_Type == "Device")
         {
-            ClearWiringGroupAndCache();
-        }
-        else
-        {
-            UpdateDeviceInformation(device);
-            currentLoadedDeviceCode = input;
-        }
-    }
+            if (!deviceDictionary.TryGetValue(input.ToLower(), out var device))
+            {
+                ClearWiringGroupAndCache();
+            }
+            else
+            {
+                UpdateDeviceInformation(device);
+                current_Search_Content = input.ToLower();
 
+            }
+        }
+        if (searchableDropDown.filter_Type == "JB/TSD")
+        {
+            if (!jBDictionary.TryGetValue(input.ToLower(), out var jB))
+            {
+                ClearWiringGroupAndCache();
+            }
+            else
+            {
+                UpdateJBInformation(jB);
+                current_Search_Content = input.ToLower();
+            }
+        }
+
+    }
     private void ClearWiringGroupAndCache()
     {
         foreach (Transform child in JB_Connection_Group.transform)
@@ -149,11 +179,11 @@ public class Dropdown_On_ValueChange : MonoBehaviour
                 Destroy(child.gameObject);
             }
         }
-        // spriteCache.Clear();
     }
 
     private void UpdateDeviceInformation(DeviceInformationModel device)
     {
+        if (!deviceInfo.gameObject.activeSelf) deviceInfo.gameObject.SetActive(true);
         prefab_Device.name = $"Scroll_Area_{device.Code}";
         code_Value_Text.text = device.Code;
         function_Value_Text.text = device.Function;
@@ -174,30 +204,43 @@ public class Dropdown_On_ValueChange : MonoBehaviour
         }
     }
 
+    private void UpdateJBInformation(JBInformationModel jB)
+    {
+        if (deviceInfo.gameObject.activeSelf) deviceInfo.gameObject.SetActive(false);
+        prefab_Device.name = $"Scroll_Area_{jB.Name}";
+        jb_Connection_Value_Text.text = $"{jB.Name}:";
+        jb_Connection_Location_Text.text = jB.Location;
+        _jbName = jb_Connection_Value_Text.text;
+        GlobalVariable_Search_Devices.jbName = _jbName;
+        if (!string.IsNullOrEmpty(_jbName))
+        {
+            ClearWiringGroupAndCache();
+            LoadDeviceSprites(null, jbInformationModel: jB);
+        }
+    }
+
     private async void LoadDeviceSprites(List<string> list_Additional_Images, JBInformationModel jbInformationModel)
     {
-        jbInformationModel.ListConnectionImages.AddRange(list_Additional_Images);
+        if (list_Additional_Images != null) jbInformationModel.ListConnectionImages.AddRange(list_Additional_Images);
         await Apply_Sprite_JB_Images(jbInformationModel.OutdoorImage, jbInformationModel.ListConnectionImages);
         scrollRect.verticalNormalizedPosition = 1f;
-        jbInformationModel.ListConnectionImages.RemoveRange(jbInformationModel.ListConnectionImages.Count - list_Additional_Images.Count, list_Additional_Images.Count);
+        if (list_Additional_Images != null) jbInformationModel.ListConnectionImages.RemoveRange(jbInformationModel.ListConnectionImages.Count - list_Additional_Images.Count, list_Additional_Images.Count);
+
     }
 
     private async Task Apply_Sprite_JB_Images(string outdoorImage, List<string> listConnectionImages)
     {
-        // Tắt các prefab ban đầu
         JB_Location_Image_Prefab.gameObject.SetActive(false);
         JB_Connection_Wiring_Image_Prefab.gameObject.SetActive(false);
 
         var tasks = new List<Task>();
 
-        // Tải hình ảnh vị trí ngoài trời nếu có
         if (!string.IsNullOrEmpty(outdoorImage))
         {
             AddButtonListener(JB_Location_Image_Prefab, outdoorImage);
             tasks.Add(LoadImageAsync(outdoorImage, JB_Location_Image_Prefab));
         }
 
-        // Tải hình ảnh kết nối
         foreach (string connectionImage in listConnectionImages)
         {
             var newImage = Instantiate(JB_Connection_Wiring_Image_Prefab, JB_Connection_Group.transform);
@@ -205,23 +248,16 @@ public class Dropdown_On_ValueChange : MonoBehaviour
             tasks.Add(LoadImageAsync(connectionImage, newImage));
         }
 
-        // Hiển thị thanh tiến trình
         await ShowProgressBar("Đang tra cứu...", "Dữ liệu đang được tải...");
-
-        // Theo dõi tiến trình tải ảnh
         await TrackProgress(tasks);
-
-        // Đặt kích thước ảnh phù hợp
         ResizeImages();
-
-        // Ẩn thanh tiến trình
         await HideProgressBar();
     }
 
     private void AddButtonListener(Image imagePrefab, string imageName)
     {
         var button = imagePrefab.GetComponent<Button>();
-        button.onClick.RemoveAllListeners(); // Đảm bảo không lặp lại listener
+        button.onClick.RemoveAllListeners();
         button.onClick.AddListener(() => open_Detail_Image.Open_Detail_Canvas(imagePrefab));
     }
 
@@ -233,7 +269,6 @@ public class Dropdown_On_ValueChange : MonoBehaviour
 
     private async Task ShowProgressBar(string title, string details)
     {
-
         Progress.Show(title, ProgressColor.Blue, true);
         Progress.SetDetailsText(details);
         await Task.Delay(100);
@@ -257,7 +292,6 @@ public class Dropdown_On_ValueChange : MonoBehaviour
             finally
             {
                 completedTasks++;
-
                 Progress.SetProgressValue((float)completedTasks / totalTasks * 100f);
             }
         });
@@ -280,7 +314,6 @@ public class Dropdown_On_ValueChange : MonoBehaviour
         await Task.Delay(200);
         Progress.Hide();
     }
-
 
     private void SetSprite(Image imageComponent, string jb_name)
     {
@@ -314,10 +347,23 @@ public class Dropdown_On_ValueChange : MonoBehaviour
         }
         instantiatedImages.Clear();
     }
+
     private void ResetResources()
     {
         searchableDropDown.inputField.onValueChanged.RemoveListener(OnInputValueChanged);
         spriteCache.Clear();
-        currentLoadedDeviceCode = null;
+        current_Search_Content = null;
+    }
+
+    private void HandleOrientationChange(ScreenOrientation newOrientation)
+    {
+        if (newOrientation == ScreenOrientation.Portrait)
+        {
+            bottom_App_Bar.SetActive(true);
+        }
+        else if (newOrientation == ScreenOrientation.LandscapeLeft || newOrientation == ScreenOrientation.LandscapeRight)
+        {
+            bottom_App_Bar.SetActive(false);
+        }
     }
 }
